+++ tv/src/main/java/ca/devmesh/seerrtv/ui/components/CustomSearchBar.kt
@@ -146,8 +146,6 @@ fun CustomSearchBar(
                                     }
                                     textFieldFocusRequester.requestFocus()
                                     return@onPreviewKeyEvent true
-                                } else {
-                                    false
                                 }
                             }
                         }